requirements/training.txt
# 训练环境依赖
# PyTorch 生态
torch>=2.0.0
torchvision>=0.15.0

# YOLO-OBB
ultralytics>=8.0.200
opencv-python>=4.8.0

# 数据处理
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0

# 模型导出
onnx>=1.14.0
onnxsim>=0.4.36

# 工具
tqdm>=4.65.0
tensorboard>=2.13.0
seaborn>=0.12.0

# 开发工具
pytest>=7.4.0
